Adding "Current Month's Releases" Playlist
It was suggested in one of the comments on the announcement post that having a continually-updating playlist that is basically an alias for the "current month"'s releases would be a good idea.
This would allow followers to only need to stay subscribed to that one playlist and not need to worry about future ones.

This required a couple technical changes.

1. Adding support for PUT requests to Curl. Minor refactor to allow any HTTP method.
2. Adding relevant methods to the SpotifyClient class to handle the new "Current" playlist.
3. Modifications to the main script runner to update both the specific-month's playlist and the current month's playlist releases.
4. Moving the function that marks a particular release as "processed" out of the Spotify client and into the main script runner. This is so that the release gets added to all the playlists it should before getting marked as finished.

Using a magic value of 'Current' for the month column in the Spotify DB to store this value. The year is set to 0 since it's just ignored.

Lastly, there needed to be a way to tell when to clear out the playlist when there was a new month.
Easiest way I could think to do that would be to leverage the script's run cadence (cron job set to every quarter hour) and check the time.
If we are on the first day of a month, the 0th hour, and the minutes are fewer than 15, we are on the first run of the new month. Which is the perfect time to clear out the playlist.

There will be a kind of weird gap between the PST-timezone start of the month when the playlist is emptied out and when the first release is added to the releases wiki. But since the month-specific playlist will still exist, I'm not worried.

I also plan on adding a "25 Most Recent Releases" playlist soon which should solve that rotating problem as well.

I tested in prod locally on my machine so I'm pretty confident this all works as-intended.

New playlist URL is here: https://open.spotify.com/playlist/7GOF5M1Nz29ALFzWIwfrdI