Add all the comments! (+ small bugfix)
Serious Code should have Serious Comments.
Used the Javadoc/PHPDoc annotation format for all of the functions in the script to make it eaiser if anyone chooses to read this code.

Speaking of, if you're reading this, feel free to file a pull request if you think this code could be better. :)

There are also two small tweaks buried in here:
1. Also move the check to see if the release is processed into the top level. Otherwise we get a bunch of warnings.
2. Small refactor on the Curl class to just use the constructor where applicable.

Nether of those should have really changed anything substantial. Tested in prod.