Adding year-long playlists
It was suggested that in addition to the monthly playlists it would be nice to have full-years. So... this adds those!
Using the magic value of 'All' for the month value in the DB for these playlists.

But yeah adding these was easy enough given the existing code.
In addition to the monthly and current-month playlists, just added a line to add to the current-year's playlists as well.
Should also handle automatically creating the new playlist when the year changes over.

Tested locally by creating yearly playlists for the previous 3.5ish years (2020 – 2017). Worked great!