Catch straggling releases from previous month
When we change over to a new month, chances are all the streaming links
from the previous month have not been added yet. So this modifies the
script slightly to update the previous month if we are still within the
first week of our new month. Made sure to account for January where the
previous month is also the previous year.

Also fixed a small bug with the $day variable since I previously forgot
to cast it to `int`... which meant my strict comparison was failing.
A+ programming right there.